Move make_date, to_char to datafusion-functions
