+++ datafusion/expr/src/built_in_function.rs
@@ -22,7 +22,6 @@ use std::fmt;
 use std::str::FromStr;
 use std::sync::{Arc, OnceLock};
 
-use crate::signature::TIMEZONE_WILDCARD;
 use crate::type_coercion::functions::data_types;
 use crate::{FuncMonotonicity, Signature, TypeSignature, Volatility};
 
@@ -228,8 +227,6 @@ pub enum BuiltinScalarFunction {
     CurrentDate,
     /// current_time
     CurrentTime,
-    /// make_date
-    MakeDate,
     /// translate
     Translate,
     /// trim
@@ -248,8 +245,6 @@ pub enum BuiltinScalarFunction {
     SubstrIndex,
     /// find_in_set
     FindInSet,
-    /// to_char
-    ToChar,
 }
 
 /// Maps the sql function name to `BuiltinScalarFunction`
@@ -388,8 +383,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::Strpos => Volatility::Immutable,
             BuiltinScalarFunction::Substr => Volatility::Immutable,
             BuiltinScalarFunction::ToHex => Volatility::Immutable,
-            BuiltinScalarFunction::ToChar => Volatility::Immutable,
-            BuiltinScalarFunction::MakeDate => Volatility::Immutable,
             BuiltinScalarFunction::Translate => Volatility::Immutable,
             BuiltinScalarFunction::Trim => Volatility::Immutable,
             BuiltinScalarFunction::Upper => Volatility::Immutable,
@@ -581,14 +574,12 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::FindInSet => {
                 utf8_to_int_type(&input_expr_types[0], "find_in_set")
             }
-            BuiltinScalarFunction::ToChar => Ok(Utf8),
             BuiltinScalarFunction::FromUnixtime => Ok(Timestamp(Second, None)),
             BuiltinScalarFunction::Now => {
                 Ok(Timestamp(Nanosecond, Some("+00:00".into())))
             }
             BuiltinScalarFunction::CurrentDate => Ok(Date32),
             BuiltinScalarFunction::CurrentTime => Ok(Time64(Nanosecond)),
-            BuiltinScalarFunction::MakeDate => Ok(Date32),
             BuiltinScalarFunction::Translate => {
                 utf8_to_str_type(&input_expr_types[0], "translate")
             }
@@ -675,7 +666,6 @@ impl BuiltinScalarFunction {
     /// Return the argument [`Signature`] supported by this function
     pub fn signature(&self) -> Signature {
         use DataType::*;
-        use TimeUnit::*;
         use TypeSignature::*;
         // note: the physical expression must accept the type returned by this function or the execution panics.
 
@@ -775,41 +765,6 @@ impl BuiltinScalarFunction {
             | BuiltinScalarFunction::Repeat
             | BuiltinScalarFunction::Right => Signature::one_of(
                 vec![Exact(vec![Utf8, Int64]), Exact(vec![LargeUtf8, Int64])],
-                self.volatility(),
-            ),
-            BuiltinScalarFunction::ToChar => Signature::one_of(
-                vec![
-                    Exact(vec![Date32, Utf8]),
-                    Exact(vec![Date64, Utf8]),
-                    Exact(vec![Time32(Millisecond), Utf8]),
-                    Exact(vec![Time32(Second), Utf8]),
-                    Exact(vec![Time64(Microsecond), Utf8]),
-                    Exact(vec![Time64(Nanosecond), Utf8]),
-                    Exact(vec![Timestamp(Second, None), Utf8]),
-                    Exact(vec![
-                        Timestamp(Second, Some(TIMEZONE_WILDCARD.into())),
-                        Utf8,
-                    ]),
-                    Exact(vec![Timestamp(Millisecond, None), Utf8]),
-                    Exact(vec![
-                        Timestamp(Millisecond, Some(TIMEZONE_WILDCARD.into())),
-                        Utf8,
-                    ]),
-                    Exact(vec![Timestamp(Microsecond, None), Utf8]),
-                    Exact(vec![
-                        Timestamp(Microsecond, Some(TIMEZONE_WILDCARD.into())),
-                        Utf8,
-                    ]),
-                    Exact(vec![Timestamp(Nanosecond, None), Utf8]),
-                    Exact(vec![
-                        Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
-                        Utf8,
-                    ]),
-                    Exact(vec![Duration(Second), Utf8]),
-                    Exact(vec![Duration(Millisecond), Utf8]),
-                    Exact(vec![Duration(Microsecond), Utf8]),
-                    Exact(vec![Duration(Nanosecond), Utf8]),
-                ],
                 self.volatility(),
             ),
             BuiltinScalarFunction::FromUnixtime => {
@@ -974,11 +929,6 @@ impl BuiltinScalarFunction {
             | BuiltinScalarFunction::CurrentTime => {
                 Signature::uniform(0, vec![], self.volatility())
             }
-            BuiltinScalarFunction::MakeDate => Signature::uniform(
-                3,
-                vec![Int32, Int64, UInt32, UInt64, Utf8],
-                self.volatility(),
-            ),
             BuiltinScalarFunction::Iszero => Signature::one_of(
                 vec![Exact(vec![Float32]), Exact(vec![Float64])],
                 self.volatility(),
@@ -1106,8 +1056,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::Now => &["now"],
             BuiltinScalarFunction::CurrentDate => &["current_date", "today"],
             BuiltinScalarFunction::CurrentTime => &["current_time"],
-            BuiltinScalarFunction::MakeDate => &["make_date"],
-            BuiltinScalarFunction::ToChar => &["to_char", "date_format"],
             BuiltinScalarFunction::FromUnixtime => &["from_unixtime"],
 
             // hashing functions
