+++ datafusion/expr/src/expr_fn.rs
@@ -800,12 +800,6 @@ nary_scalar_expr!(
 
 // date functions
 scalar_expr!(
-    ToChar,
-    to_char,
-    datetime format,
-    "converts a date, time, timestamp or duration to a string based on the provided format"
-);
-scalar_expr!(
     FromUnixtime,
     from_unixtime,
     unixtime,
@@ -814,7 +808,6 @@ scalar_expr!(
 scalar_expr!(CurrentDate, current_date, ,"returns current UTC date as a [`DataType::Date32`] value");
 scalar_expr!(Now, now, ,"returns current timestamp in nanoseconds, using the same value for all instances of now() in same statement");
 scalar_expr!(CurrentTime, current_time, , "returns current UTC time as a [`DataType::Time64`] value");
-scalar_expr!(MakeDate, make_date, year month day, "make a date from year, month and day component parts");
 scalar_expr!(Nanvl, nanvl, x y, "returns x if x is not NaN otherwise returns y");
 scalar_expr!(
     Iszero,
