+++ datafusion/functions/Cargo.toml
@@ -51,6 +51,7 @@ path = "src/lib.rs"
 [dependencies]
 arrow = { workspace = true }
 arrow-array = { workspace = true }
+arrow-schema = { workspace = true }
 base64 = { version = "0.22", optional = true }
 chrono = { workspace = true }
 datafusion-common = { workspace = true }
@@ -75,3 +76,11 @@ name = "to_timestamp"
 [[bench]]
 harness = false
 name = "regx"
+
+[[bench]]
+harness = false
+name = "make_date"
+
+[[bench]]
+harness = false
+name = "to_char"
