+++ datafusion/physical-expr/src/functions.rs
@@ -445,8 +445,6 @@ pub fn create_physical_fun(
                 execution_props.query_execution_start_time,
             ))
         }
-        BuiltinScalarFunction::MakeDate => Arc::new(datetime_expressions::make_date),
-        BuiltinScalarFunction::ToChar => Arc::new(datetime_expressions::to_char),
         BuiltinScalarFunction::FromUnixtime => {
             Arc::new(datetime_expressions::from_unixtime_invoke)
         }
