+++ datafusion/proto/proto/datafusion.proto
@@ -680,10 +680,10 @@ enum ScalarFunction {
   ArrayResize = 130;
   EndsWith = 131;
   /// 132 was InStr
-  MakeDate = 133;
+  /// 133 was MakeDate
   ArrayReverse = 134;
   /// 135 is RegexpLike
-  ToChar = 136;
+  /// 136 was ToChar
   /// 137 was ToDate
   /// 138 was ToUnixtime
 }
